Add autopkgtest support
The default autopkgtest-pkg-perl tests don't really work with this package.
The build system generates t/TEST, which uses ModPerl::TestRun to
run the tests under t/ subdirectories. Also, compiling the .pm files
shipped in this package will generally fail because they need to be
used in an Apache environment.

So we roll our own infrastructure and opt out of autopkgtest-pkg-perl.
The 'mytest' script generates a minimal t/TEST and calls it with a
few options that are apparently needed when we're not in a full build
directory.

Somewhat surprisingly, this works for now. It may need amending in the
future, but mod_perl2 releases aren't too frequent nowadays.